microbenchmarks/requirements.txt
torch
cupti-python
scipy